requirements.txt
google-cloud-speech
pyaudio
Pillow
pytesseract
openai